It's done!
Iteration loops work now rather nice. Since I've moved to SVM iteration
inside the k iteration time should have improved as well.
BUT: I fear my algorithms are not working correctly, because pretty soon
there is no more changement in the accuracy. Or is it supposed to be
like that?

Anyway, all thats left to do is cleaning up and evaluation.