Nice 'n Shiny!
Everything seems to work fine. Evaluation took about 80 minutes.
Added plot and file where you can see that for k >= 26 we have an
accuracy of 90% and for k >= 46 we have an almost constant accuracy
of 91%.
I will let it run again, but only for k's between 40 and 50 to have
an excat result.